+++ plx--1.1.1--1.2.sql
@@ -17,3 +17,21 @@ CREATE TRUSTED LANGUAGE plxplsql
 	VALIDATOR plx_plsql_validator;
 
 COMMENT ON LANGUAGE plxplsql IS 'plx Oracle PL/SQL dialect (transpiles to plpgsql)';
+
+/* TypeScript dialect validator + inline handler live in plx.so */
+CREATE FUNCTION plx_ts_validator(oid)
+	RETURNS void
+	AS '$libdir/plx', 'plx_ts_validator'
+	LANGUAGE C STRICT;
+
+CREATE FUNCTION plx_ts_inline_handler(internal)
+	RETURNS void
+	AS '$libdir/plx', 'plx_ts_inline_handler'
+	LANGUAGE C;
+
+CREATE TRUSTED LANGUAGE plxts
+	HANDLER plx_call_handler
+	INLINE plx_ts_inline_handler
+	VALIDATOR plx_ts_validator;
+
+COMMENT ON LANGUAGE plxts IS 'plx TypeScript dialect (transpiles to plpgsql)';